+++ src/parse.cpp
@@ -50,10 +50,12 @@ std::vector<int> count_fields_(List sourceSpec, List tokenizerSpec, int n_max) {
 
 // [[Rcpp::export]]
 RObject read_header_(List sourceSpec, List tokenizerSpec, List locale_) {
+  Warnings warnings;
   LocaleInfo locale(locale_);
   SourcePtr source = Source::create(sourceSpec);
   TokenizerPtr tokenizer = Tokenizer::create(tokenizerSpec);
   tokenizer->tokenize(source->begin(), source->end());
+  tokenizer->setWarnings(&warnings);
 
   CollectorCharacter out(&locale.encoder_);